renamed `common_utility` to `utility` as it was redeploying all modules because of regex match of `common`
