+++ utility/application/handlers/upload_handlers.py
@@ -5,17 +5,17 @@ from common.constant import StatusCode
 from common.exception_handler import exception_handler
 from common.logger import get_logger
 from common.utils import generate_lambda_response, validate_dict
-from common_utility.application.upload_service import UploadService
-from common_utility.config import SLACK_HOOK, NETWORK_ID
-from common_utility.constants import UPLOAD_TYPE_DETAILS, ALLOWED_CONTENT_TYPE, FILE_EXTENSION, TEMP_FILE_DIR
-from common_utility.exceptions import EXCEPTIONS, InvalidContentType, BadRequestException
+from utility.application.upload_service import UploadService
+from utility.config import SLACK_HOOK, NETWORK_ID
+from utility.constants import UPLOAD_TYPE_DETAILS, ALLOWED_CONTENT_TYPE, FILE_EXTENSION, TEMP_FILE_DIR
+from utility.exceptions import EXCEPTIONS, InvalidContentType, BadRequestException
 
 logger = get_logger(__name__)
 
 
 @exception_handler(SLACK_HOOK=SLACK_HOOK, NETWORK_ID=NETWORK_ID, logger=logger, EXCEPTIONS=EXCEPTIONS)
 def upload_file(event, context):
-    content_type = event["header"]["Content-Type"]
+    content_type = event["headers"]["Content-Type"]
     username = event["requestContext"]["authorizer"]["claims"]["email"]
     query_string_parameter = event["queryStringParameters"]
