+++ utility/application/upload_service.py
@@ -2,8 +2,8 @@ from common.boto_utils import BotoUtils
 from common.exceptions import BadRequestException
 from common.logger import get_logger
 from common.utils import date_time_for_filename
-from common_utility.config import REGION_NAME
-from common_utility.constants import UPLOAD_TYPE_DETAILS, UploadType
+from utility.config import REGION_NAME
+from utility.constants import UPLOAD_TYPE_DETAILS, UploadType
 
 logger = get_logger(__name__)
